requirements.txt
celery
Django
dnspython
gevent
python-nmap
redis
requests
urllib3
IPy
eventlet
